Absolute without convert

+++ 0x02-functions_nested_loops/6-abs.c
@@ -13,11 +13,11 @@ int _abs(int i)
 	if (i < '0')
 	{
 		val = (i * -1);
-		_putchar(val);
+		_putchar((int)val);
 	}
 	else
 	{
-		_putchar(i);
+		_putchar((int)i);
 	}
 
 	return (0);
